🛠️ Fixed warnings -Wall (kept recursion ones)

+++ games/hanged/main.cpp
@@ -38,14 +38,14 @@ void DrawHangman(int fail, vector<char> wordByCharGuessed, vector<char> alreadyT
          << endl;
 
     string wordByCharGuessedString = "";
-    for (int i = 0; i < wordByCharGuessed.size(); i++)
+    for (unsigned long i = 0; i < wordByCharGuessed.size(); i++)
     {
         wordByCharGuessedString += wordByCharGuessed[i];
         wordByCharGuessedString += " ";
     }
 
     string alreadyTestedString = "";
-    for (int i = 0; i < alreadyTested.size(); i++)
+    for (unsigned long i = 0; i < alreadyTested.size(); i++)
         alreadyTestedString += alreadyTested[i];
 
     cout << "     "
@@ -75,7 +75,7 @@ extern void Hangman()
         wordByCharGuessed;      // Word by char guessed
     vector<char> alreadyTested; // Already tested letters
 
-    for (int i = 0; i < word.length(); i++)
+    for (unsigned long i = 0; i < word.length(); i++)
     {
         wordByChar.push_back(word[i]);
         wordByCharGuessed.push_back('_');
@@ -91,7 +91,7 @@ extern void Hangman()
         if (test == 3) // Handle ^C
             return;
 
-        for (int i = 0; i < alreadyTested.size(); i++)
+        for (unsigned long i = 0; i < alreadyTested.size(); i++)
         {
             if (alreadyTested[i] == test)
                 alreadyTestedBool = true;
@@ -104,7 +104,7 @@ extern void Hangman()
 
         bool found = false;
 
-        for (int i = 0; i < word.length(); i++)
+        for (unsigned long i = 0; i < word.length(); i++)
         {
             if (wordByChar[i] == test)
             {
@@ -119,7 +119,7 @@ extern void Hangman()
         DrawHangman(fail, wordByCharGuessed, alreadyTested);
         bool allGuessed = true;
 
-        for (int i = 0; i < word.length(); i++)
+        for (unsigned long i = 0; i < word.length(); i++)
         {
             if (wordByCharGuessed[i] == '_')
                 allGuessed = false;
